requirements.txt
qdrant-client
plivo==4.58.6
requests==2.31.0
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
httpx
python-multipart
openai
redis
boto3
aiohttp
motor==3.7.0
pymongo==4.11.3
google-genai
pydantic-settings
aiofiles
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
langchain-qdrant>=0.1.4
langchain-openai>=0.3.0
openpyxl==3.1.0
xlrd==2.0.1
google
asyncio
PyMuPDF
psutil
langchain_experimental
PyPDF2